moved numberToString function and adds unit tests
